提交：1、修改目录；2、添加podspec

+++ Example/UnlockByID/ViewController.m
@@ -30,7 +30,7 @@
 - (void)checkID
 {
     [[UnlockByIDUtils shareManager] showVerityWithAllowPassword:YES result:^(BOOL isSuc, UnlockByIDUtilsState state) {
-        NSLog(@"%d",state);
+        NSLog(@"%ld",(long)state);
     }];
 }
